+++ api/v2/urls.py
@@ -8,7 +8,7 @@ from groups import api as groups_api
 from reader import api as reader_api
 from users import api as users_api
 
-from .views import openapi, redoc_redirect, swagger_redirect
+from .views import openapi, rapidoc, redoc_redirect, swagger_redirect
 
 #: The API router
 router = SimpleRouter(trailing_slash=False)
@@ -32,6 +32,7 @@ urlpatterns = [
     # HACK: move profile operations to the main endpoint
     path('profile', users_api.ProfileViewSet.as_view()),
     path('openapi.json', openapi, name='schema'),
+    path('docs/', rapidoc, name='rapidoc'),
     path('redoc/', redoc_redirect, name='redoc'),
     path('swagger/', swagger_redirect, name='swagger'),
 ]