+++ groups/api.py
@@ -1,5 +1,6 @@
 """API viewsets for the groups app."""
 
+from rest_framework.parsers import MultiPartParser
 from rest_framework.viewsets import ModelViewSet
 
 from api.v2.mixins import CORSMixin
@@ -15,13 +16,14 @@ class GroupViewSet(CORSMixin, ModelViewSet):
     * list: List groups.
     * read: View a certain group.
     * create: Create a new group.
-    * update: Edit the given author.
-    * patch: Patch the given group.
+    * update: Update the given author.
+    * patch: Edit the given group.
     * delete: Delete the given group.
     """
     schema = OpenAPISchema(tags=('groups',))
     queryset = models.Group.objects.all()
     serializer_class = serializers.GroupSerializer
+    parser_classes = (MultiPartParser,)
 
 
 __all__ = ['GroupViewSet']